requirements.txt
attrs==21.2.0
black==21.9b0
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
Flask==2.0.1
Flask-Testing==0.8.1
idna==3.2
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mypy-extensions==0.4.3
packaging==21.0
pathspec==0.9.0
platformdirs==2.4.0
pluggy==1.0.0
psycopg2-binary==2.9.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
regex==2021.9.30
requests==2.26.0
simplejson==3.17.5
timeout-decorator==0.5.0
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.1
